dao implementado

+++ src/dao/cart.dao.js
@@ -1,7 +1,11 @@
 import CartModel from "../models/cart.model.js";
+import mongoDB from "../config/mongoose.config.js";
 
 class CartDao {
     async findById(id) {
+        if (!mongoDB.isValidId(id)) {
+            return "ID no válido";
+        }
         return await CartModel.findById(id);
     }
 
@@ -9,16 +13,26 @@ class CartDao {
         return await CartModel.findOne(query);
     }
 
-    async save(userData) {
-        const user = new CartModel(userData);
-        return await user.save();
+    async find() {
+        return await CartModel.find().populate("products").lean();
     }
 
-    async update(id, userData) {
-        return await CartModel.findByIdAndUpdate(id, userData);
+    async save(cartData) {
+        const cart = new CartModel(cartData);
+        return await cart.save();
+    }
+
+    async update(id, cartData) {
+        if (!mongoDB.isValidId(id)) {
+            return "ID no válido";
+        }
+        return await CartModel.findByIdAndUpdate(id, cartData, { new: true });
     }
 
     async delete(id) {
+        if (!mongoDB.isValidId(id)) {
+            return "ID no válido";
+        }
         return await CartModel.findByIdAndDelete(id);
     }
 }
