+++ src/dao/product.dao.js
@@ -1,8 +1,23 @@
 import ProductModel from "../models/product.model.js";
+import mongoDB from "../config/mongoose.config.js";
 
 class ProductDao {
     async findById(id) {
+        if (!mongoDB.isValidId(id)) {
+            return "ID no válido";
+        }
         return await ProductModel.findById(id);
+    }
+
+    async findByIdAndUpdate(id, productData) {
+        if (!mongoDB.isValidId(id)) {
+            return "ID no válido";
+        }
+        return await ProductModel.findByIdAndUpdate(id, productData);
+    }
+
+    async find() {
+        return await ProductModel.find();
     }
 
     async findOne(query) {
@@ -15,10 +30,16 @@ class ProductDao {
     }
 
     async update(id, productData) {
+        if (!mongoDB.isValidId(id)) {
+            return "ID no válido";
+        }
         return await ProductModel.findByIdAndUpdate(id, productData);
     }
 
     async delete(id) {
+        if (!mongoDB.isValidId(id)) {
+            return "ID no válido";
+        }
         return await ProductModel.findByIdAndDelete(id);
     }
 
@@ -30,6 +51,10 @@ class ProductDao {
             lean: true,
             pagination: true,
         });
+    }
+
+    async explain(filters) {
+        return await ProductModel.find(filters).explain();
     }
 }
