+++ src/dao/user.dao.js
@@ -1,7 +1,11 @@
 import UserModel from "../models/user.model.js";
+import mongoDB from "../config/mongoose.config.js";
 
 class UserDao {
     async findById(id) {
+        if (!mongoDB.isValidId(id)) {
+            return "ID no válido";
+        }
         return await UserModel.findById(id);
     }
 
@@ -15,10 +19,16 @@ class UserDao {
     }
 
     async update(id, userData) {
+        if (!mongoDB.isValidId(id)) {
+            return "ID no válido";
+        }
         return await UserModel.findByIdAndUpdate(id, userData);
     }
 
     async delete(id) {
+        if (!mongoDB.isValidId(id)) {
+            return "ID no válido";
+        }
         return await UserModel.findByIdAndDelete(id);
     }
 }
